requirements.txt
pyyaml
unix-ar
zstandard